Add shifting of types from common to invoice specific package
