+++ packages/invoices/src/OutboundShipment/DetailView/api.ts
@@ -2,7 +2,6 @@ import {
   gql,
   Name,
   SortBy,
-  Invoice,
   GraphQLClient,
   getSdk,
   NameSortFieldInput,
@@ -22,7 +21,7 @@ import {
   UpdateOutboundShipmentLineInput,
 } from '@openmsupply-client/common';
 import { Environment } from '@openmsupply-client/config';
-import { OutboundShipment, OutboundShipmentRow } from '../../types';
+import { OutboundShipment, OutboundShipmentRow, Invoice } from '../../types';
 import { flattenSummaryItems } from '../../utils';
 
 const client = new GraphQLClient(Environment.API_URL);
