+++ packages/invoices/src/types.ts
@@ -1,11 +1,36 @@
 import {
-  Invoice,
   Column,
   InvoiceLine,
   StockLine,
-  OutboundShipmentStatus,
-  InboundShipmentStatus,
+  InvoiceNode,
+  InvoiceNodeStatus,
+  DomainObject,
+  InvoicePricingNode,
+  Name,
 } from '@openmsupply-client/common';
+
+export type OutboundShipmentStatus =
+  | InvoiceNodeStatus.Draft
+  | InvoiceNodeStatus.Allocated
+  | InvoiceNodeStatus.Picked
+  | InvoiceNodeStatus.Shipped
+  | InvoiceNodeStatus.Delivered;
+
+export type InboundShipmentStatus =
+  | InvoiceNodeStatus.Picked
+  | 'NEW'
+  | InvoiceNodeStatus.Shipped
+  | InvoiceNodeStatus.Delivered
+  | 'VERIFIED';
+
+export interface Invoice
+  extends Omit<InvoiceNode, 'lines' | 'status' | 'otherParty'>,
+    DomainObject {
+  status: InvoiceNodeStatus | 'NEW' | 'VERIFIED';
+  otherParty?: Name;
+  lines: InvoiceLine[];
+  pricing: InvoicePricingNode;
+}
 
 export interface BatchRow extends StockLine {
   numberOfPacks: number;
